Append details on parameters to expanded function docstrings.
Test runners will occassionaly use the docstring of a test method
as a means of reporting its execution. Because the docstring is
simply copied from the template method to the parameterized
method, it isn't very useful in distinguishing between tests.

We now detect the names of the parameters of the function and
the arguments that are being passed as parameters and append
that information to the end of the first line of the docstring.

The generated docstrings might not make perfect sense, so the
user can also specify testcase_func_doc to override docstring
generation with information that makes more sense.

Fixes #21