Issue #90: Port setup.py to Python 3.9
Avoid "U" mode in open(). This flag was ignored since Python 3.0 and
now raises an error in Python 3.9:
https://bugs.python.org/issue37330

On Python 2, setup.py no longer replaces DOS (\r\n) and Mac (\r) with
Unix newline (\n). But it is not an issue in practice, since
README.rst uses Unix newline format (\n).

Fix also a ResourceWarning: use "with" context manager to ensure that
the file is closed. The warning was displayed when using
"python3 -Wd setup.py (...)".